Moved the check for 32-bit before the check for 64-bit
If an architecture is mistakenly detected as 32-bit due to a bug,
there’s not much harm done. If an architecture is mistakenly detected
as 64-bit, however, the download won’t run. Therefore, it makes sense
to err on the side of caution and put the check for a 32-bit
architecture first.