added nojs.html (NO JAVASCRIPT PAGE)
Links do not work so please do not click them.

There are a few ways we can implement this:

a) load as div that is hidden and other page is shown using javascript.
b) have user click link when js is not present…

probably other ways though, if anyone has ideas please share…