Made download start on the thank-you page
This allows the back button to function properly.

It also has the benefit of allowing a user to pass a direct link to a
friend, removing the need for someone to even go through the drop
downs. This setup is a good balance between a direct link (where there
is no information provided) and the drop-downs (which a friend may have
already filled out when linking).